+++ backend/molgenis-emx2-fairmapper/src/main/java/org/molgenis/emx2/fairmapper/preprocessing/RdfPreProcessor.java
@@ -1,15 +1,15 @@
-package org.molgenis.emx2.fairmapper.enrich;
+package org.molgenis.emx2.fairmapper.preprocessing;
 
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 
 /**
- * Optional enrichment step in the DCAT harvesting pipeline, applied after RDF extraction and before
+ * Optional pre-processing step in the DCAT harvesting pipeline, applied after RDF extraction and before
  * the SPARQL SELECT transformation.
  *
  * <p>Implementations mutate the repository by adding triples that can be derived from the existing
  * graph, making the data richer or more uniform for downstream queries.
  */
-public interface RdfEnricher {
+public interface RdfPreProcessor {
 
   /**
    * Adds derived triples to {@code repository}.
@@ -17,5 +17,5 @@ public interface RdfEnricher {
    * <p>Implementations may read from and write to the same repository. All mutations must be
    * committed before this method returns so that subsequent pipeline steps see the enriched graph.
    */
-  void enrich(SailRepository repository);
+  void process(SailRepository repository);
 }
